Let `ChannelSigner` build the inputs of the commit tx
